requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
behave==1.2.6
certifi==2025.1.31
cffi==1.17.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
